clean

+++ classes/register.class.php
@@ -57,9 +57,20 @@ class RegisterUser
         return false;
     }
 
+    private function emailExists()
+    {
+        foreach ($this->stored_users as $user) {
+            if ($this->email == $user['email']) {
+                $this->invalid_feedback = "email already taken, please choose a differnet one.";
+                return true;
+            }
+        }
+        return false;
+    }
+
     private function insertUser()
     {
-        if ($this->usernameExists() == FALSE) {
+        if ($this->usernameExists() == FALSE && $this->emailExists() == FALSE) {
             array_push($this->stored_users, $this->new_user);
 
             if (file_put_contents($this->storage, json_encode($this->stored_users, JSON_PRETTY_PRINT))) {
